Wire preference controller into a lifecycle
Since we want to observe the action setting so we can update the summary, wire the preference controller into a lifecycle observer so that we can clean up any resources that we use according to the lifecycle of the Settings app. This is also used to update the checked value of the preference should it change while the user is changing the action settings.

Change-Id: I355b9a50ca2346b55f7455d5f1fb9ea574cdc385